Refactor: rename PageShare to PageExport and move export methods there
